+++ src/transport/mock.rs
@@ -92,6 +92,10 @@ impl Transport for MockTransport {
         self.sent.lock().unwrap().push(data.to_vec());
         Ok(())
     }
+
+    fn as_any_mut(&mut self) -> Option<&mut dyn std::any::Any> {
+        Some(self)
+    }
 }
 
 #[cfg(test)]